[update]move type files under the types directory
型定義用のファイルをsrc/types/配下へ移動、各コンポーネントのimportパスも修正
