+++ src/components/Navbar.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link, useLocation } from 'react-router-dom';
 import styled from 'styled-components';
 import media from 'styled-media-query';
-import { LinkList } from '../Types';
+import { LinkList } from '../types';
 
 type Props = {
   links: LinkList
